Bug: Fixed issue with time slider causing integer overflow on the server. It had to send seconds, not milliseconds.

+++ app/src/main/java/com/example/strawberry_app/network/protocol/OutgoingMessage.kt
@@ -76,7 +76,7 @@ sealed class OutgoingMessage {
 
     @Serializable
     @SerialName("seek-to")
-    data class SeekTo(val seekTo: Int): OutgoingMessage()
+    data class SeekTo(val seekTo: Long): OutgoingMessage()
 
     @Serializable
     @SerialName("send-active-playlist")
